# contracts/openapi.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ProjectClaim": {
    "properties": {
     "apn": {
      "pattern": "^\\d{10}$",
      "title": "Apn",
      "type": "string"
     },
     "ended_at": {
      "anyOf": [
       {
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ended At"
     },
     "evidence_url": {
      "format": "uri",
      "maxLength": 2083,
      "minLength": 1,
      "title": "Evidence Url",
      "type": "string"
     },
     "license_number": {
      "anyOf": [
       {
        "maxLength": 100,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "License Number"
     },
     "organization": {
      "maxLength": 160,
      "minLength": 2,
      "title": "Organization",
      "type": "string"
     },
     "role": {
      "enum": [
       "owner",
       "general_contractor",
       "framing",
       "architect",
       "engineer",
       "real_estate_agent",
       "other_trade"
      ],
      "title": "Role",
      "type": "string"
     },
     "scope": {
      "maxLength": 1500,
      "minLength": 10,
      "title": "Scope",
      "type": "string"
     },
     "sqft": {
      "anyOf": [
       {
        "exclusiveMinimum": 0.0,
        "maximum": 1000000.0,
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sqft"
     },
     "started_at": {
      "anyOf": [
       {
        "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Started At"
     }
    },
    "required": [
     "apn",
     "organization",
     "role",
     "scope",
     "evidence_url"
    ],
    "title": "ProjectClaim",
    "type": "object"
   },
   "PropertyDetail": {
    "properties": {
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "apn": {
      "title": "Apn",
      "type": "string"
     },
     "center": {
      "anyOf": [
       {
        "items": {
         "type": "number"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Center"
     },
     "conflict_count": {
      "title": "Conflict Count",
      "type": "integer"
     },
     "identity": {
      "additionalProperties": true,
      "title": "Identity",
      "type": "object"
     },
     "jurisdiction": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Jurisdiction"
     },
     "lane_signals": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Lane Signals",
      "type": "object"
     },
     "last_evidence_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Evidence Date"
     },
     "milestones": {
      "additionalProperties": {
       "type": "boolean"
      },
      "title": "Milestones",
      "type": "object"
     },
     "neighborhood": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Neighborhood"
     },
     "observation_count": {
      "title": "Observation Count",
      "type": "integer"
     },
     "pre_fire": {
      "additionalProperties": true,
      "title": "Pre Fire",
      "type": "object"
     },
     "projection_policy_version": {
      "title": "Projection Policy Version",
      "type": "string"
     },
     "property_id": {
      "title": "Property Id",
      "type": "string"
     },
     "release_id": {
      "title": "Release Id",
      "type": "string"
     },
     "snapshot_id": {
      "title": "Snapshot Id",
      "type": "string"
     }
    },
    "required": [
     "property_id",
     "apn",
     "address",
     "neighborhood",
     "jurisdiction",
     "identity",
     "pre_fire",
     "lane_signals",
     "milestones",
     "projection_policy_version",
     "observation_count",
     "conflict_count",
     "last_evidence_date",
     "center",
     "release_id",
     "snapshot_id"
    ],
    "title": "PropertyDetail",
    "type": "object"
   },
   "PropertyPage": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/PropertySummary"
      },
      "title": "Items",
      "type": "array"
     },
     "next_cursor": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Cursor"
     },
     "release_id": {
      "title": "Release Id",
      "type": "string"
     },
     "total": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Total"
     }
    },
    "required": [
     "release_id",
     "items",
     "next_cursor"
    ],
    "title": "PropertyPage",
    "type": "object"
   },
   "PropertySummary": {
    "properties": {
     "address": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Address"
     },
     "apn": {
      "title": "Apn",
      "type": "string"
     },
     "conflict_count": {
      "title": "Conflict Count",
      "type": "integer"
     },
     "jurisdiction": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Jurisdiction"
     },
     "lane_signals": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Lane Signals",
      "type": "object"
     },
     "last_evidence_date": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Evidence Date"
     },
     "milestones": {
      "additionalProperties": {
       "type": "boolean"
      },
      "title": "Milestones",
      "type": "object"
     },
     "neighborhood": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Neighborhood"
     },
     "property_id": {
      "title": "Property Id",
      "type": "string"
     }
    },
    "required": [
     "property_id",
     "apn",
     "address",
     "neighborhood",
     "jurisdiction",
     "lane_signals",
     "milestones",
     "last_evidence_date",
     "conflict_count"
    ],
    "title": "PropertySummary",
    "type": "object"
   },
...
   "TaskUpdate": {
    "properties": {
     "evidence_url": {
      "anyOf": [
       {
        "format": "uri",
        "maxLength": 2083,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Evidence Url"
     },
     "note": {
      "maxLength": 2000,
      "minLength": 5,
      "title": "Note",
      "type": "string"
     },
     "status": {
      "enum": [
       "queued",
       "in_progress",
       "needs_review"
      ],
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "status",
     "note"
    ],
    "title": "TaskUpdate",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "description": "Versioned public read API for Pacific Palisades recovery evidence. Resolve /v1/releases/current once, then use release-qualified routes.",
  "title": "OpenPali Recovery Intelligence API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/v1/evidence/releases/{release_id}/artifacts/{name}": {
   "get": {
    "operationId": "artifact_v1_evidence_releases__release_id__artifacts__name__get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Artifact",
    "tags": [
     "evidence workspace"
    ]
   }
  },
  "/v1/evidence/releases/{release_id}/claims": {
   "post": {
    "operationId": "claim_v1_evidence_releases__release_id__claims_post",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectClaim"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Claim",
    "tags": [
     "evidence workspace"
    ]
   }
  },
  "/v1/evidence/releases/{release_id}/properties": {
   "get": {
    "operationId": "properties_v1_evidence_releases__release_id__properties_get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "default": "",
       "maxLength": 120,
       "title": "Q",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "stage",
      "required": false,
      "schema": {
       "default": "",
       "title": "Stage",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 500,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Properties",
    "tags": [
     "evidence workspace"
    ]
   }
  },
  "/v1/evidence/releases/{release_id}/properties/{apn}": {
   "get": {
    "operationId": "detail_v1_evidence_releases__release_id__properties__apn__get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "apn",
      "required": true,
      "schema": {
       "title": "Apn",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Detail",
    "tags": [
     "evidence workspace"
    ]
   }
  },
  "/v1/evidence/releases/{release_id}/properties/{apn}/claims": {
   "get": {
    "operationId": "claims_v1_evidence_releases__release_id__properties__apn__claims_get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "apn",
      "required": true,
      "schema": {
       "title": "Apn",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Claims",
    "tags": [
     "evidence workspace"
    ]
   }
  },
  "/v1/evidence/releases/{release_id}/tasks": {
   "get": {
    "operationId": "tasks_v1_evidence_releases__release_id__tasks_get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "kind",
      "required": false,
      "schema": {
       "default": "",
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "offset",
      "required": false,
      "schema": {
       "default": 0,
       "minimum": 0,
       "title": "Offset",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Tasks",
    "tags": [
     "evidence workspace"
    ]
   }
  },
  "/v1/evidence/releases/{release_id}/tasks/{task_id}": {
   "post": {
    "operationId": "update_task_v1_evidence_releases__release_id__tasks__task_id__post",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "task_id",
      "required": true,
      "schema": {
       "title": "Task Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TaskUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Update Task",
    "tags": [
     "evidence workspace"
    ]
   }
  },
  "/v1/evidence/releases/{release_id}/tasks/{task_id}/history": {
   "get": {
    "operationId": "task_history_v1_evidence_releases__release_id__tasks__task_id__history_get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "task_id",
      "required": true,
      "schema": {
       "title": "Task Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Task History",
    "tags": [
     "evidence workspace"
    ]
   }
  },
...
  "/v1/releases/{release_id}/properties": {
   "get": {
    "operationId": "search_properties_v1_releases__release_id__properties_get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     },
     {
      "description": "address or APN fragment",
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 120,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "address or APN fragment",
       "title": "Q"
      }
     },
     {
      "description": "minLon,minLat,maxLon,maxLat",
      "in": "query",
      "name": "bbox",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "minLon,minLat,maxLon,maxLat",
       "title": "Bbox"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 25,
       "maximum": 100,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PropertyPage"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Search Properties",
    "tags": [
     "releases"
    ]
   }
  },
  "/v1/releases/{release_id}/properties/{property_id}": {
   "get": {
    "operationId": "property_detail_v1_releases__release_id__properties__property_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "property_id",
      "required": true,
      "schema": {
       "title": "Property Id",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "title": "Release Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PropertyDetail"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Property Detail",
    "tags": [
     "releases"
    ]
   }
  },
...
 }
}